requirements.txt
google.generativeai
flask
jsonify